requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.1
python-dotenv==1.0.1
pydantic==2.9.2
email-validator>=2.2.0
requests>=2.31.0
